fix(api): remove cache from account token inventory
