+++ apps/api/src/routes/index.ts
@@ -12,7 +12,6 @@ import keys from '#routes/keys';
 import kitwallet from '#routes/kitwallet';
 import legacy from '#routes/legacy';
 import nfts from '#routes/nfts';
-import node from '#routes/node';
 import search from '#routes/search';
 import stats from '#routes/stats';
 import sync from '#routes/sync';
@@ -33,7 +32,6 @@ const routes = () => {
   kitwallet(app);
   legacy(app);
   nfts(app);
-  node(app);
   search(app);
   stats(app);
   sync(app);
